Moves Exceptions, Randoms and RandomRule to neo4j-common
