+++ community/common/src/main/java/org/neo4j/helpers/Exceptions.java
