+++ community/consistency-check/pom.xml
@@ -126,5 +126,12 @@ the relevant Commercial Agreement.
       <artifactId>neo4j-io</artifactId>
       <version>${neo4j.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.neo4j</groupId>
+      <artifactId>neo4j-common</artifactId>
+      <version>${project.version}</version>
+      <type>test-jar</type>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 </project>
